armintas: Restructured the project
