+++ .gitignore
@@ -1,0 +1,2 @@
+models/__pycache__
+__pycache__
